+++ front-end/src/sections/shortcuts/add-shortcut-dialog/add-shortcut-dialog.tsx
@@ -1,6 +1,8 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import "./add-shortcut-dialog.css";
 import Button from "../../../widgets/button/button";
+import Spacer from "../../../widgets/spacer";
+import Alert from "../../../widgets/alert/alert";
 
 interface WebsiteInfo {
     title?: string | null;
@@ -9,6 +11,7 @@ interface WebsiteInfo {
 
 type Props = {
     onCreate: (icon: string, title: string, url: string) => void;
+    onClose: () => void;
 };
 
 export default function AddShortcutDialog(props: Props) {
@@ -19,12 +22,13 @@ export default function AddShortcutDialog(props: Props) {
 
     const [validating, setValidating] = useState(false);
     const [validated, setValidated] = useState(false);
+    const [error, setError] = useState<string | undefined>(undefined);
 
     // Methods
     const validate = () => {
-        console.log("Validating...", validated);
         if (validated) {
             if (title === "" || url === "") {
+                setError("Title and URL are required");
                 return false;
             }
 
@@ -43,6 +47,7 @@ export default function AddShortcutDialog(props: Props) {
         setValidating(true);
         if (url === "") {
             setValidating(false);
+            setError("URL is required");
             return false;
         }
 
@@ -50,11 +55,18 @@ export default function AddShortcutDialog(props: Props) {
         const urlRegex = new RegExp("^(http|https)://", "i");
         if (!urlRegex.test(url)) {
             setValidating(false);
+            setError("Invalid URL");
             return false;
         }
 
         // Get website info
         getWebsiteInfo(url).then((info) => {
+            if (!info) {
+                setError("Failed to fetch website info");
+                setValidating(false);
+                return false;
+            }
+
             setTitle(info.title || "");
             setIcon(info.favicon);
         });
@@ -85,7 +97,7 @@ export default function AddShortcutDialog(props: Props) {
         return title.charAt(0).toUpperCase() + title.slice(1);
     }
 
-    const getWebsiteInfo = async (url: string): Promise<WebsiteInfo> => {
+    const getWebsiteInfo = async (url: string): Promise<WebsiteInfo | undefined> => {
         try {
             const faviconUrl = faviconURL(url);
             const hostname = await getTitle(url);
@@ -93,9 +105,29 @@ export default function AddShortcutDialog(props: Props) {
             return { title: hostname, favicon: faviconUrl };
         } catch (error) {
             console.error("Error fetching website data:", error);
-            return { title: "Error fetching title", favicon: "" };
+            return undefined;
         }
     };
+
+    // Listeners
+    const handleKeyDown = (e: KeyboardEvent) => {
+        if (e.key === "Enter") {
+            validate();
+        }
+
+        // Close dialog
+        if (e.key === "Escape") {
+            props.onClose();
+        }
+    };
+
+    // Effects
+    useEffect(() => {
+        window.addEventListener("keydown", handleKeyDown);
+        return () => {
+            window.removeEventListener("keydown", handleKeyDown);
+        };
+    });
 
     // Render
     return (
@@ -105,20 +137,23 @@ export default function AddShortcutDialog(props: Props) {
                     <img src={icon} />
                 </div>
 
-                <div className="field">
-                    <label>Title</label>
-                    <input
-                        type="text"
-                        value={title}
-                        onChange={(e) => setTitle(e.target.value)}
-                        className={validating && !title ? "invalid" : ""}
-                    />
-                </div>
+                <Spacer height={20} />
+                {validated && (
+                    <div className="field">
+                        <input
+                            type="text"
+                            placeholder="Title"
+                            value={title}
+                            onChange={(e) => setTitle(e.target.value)}
+                            className={validating && !title ? "invalid" : ""}
+                        />
+                    </div>
+                )}
 
                 <div className="field">
-                    <label>URL</label>
                     <input
                         type="text"
+                        placeholder="URL"
                         value={url}
                         onChange={(e) => {
                             setValidated(false);
@@ -128,13 +163,39 @@ export default function AddShortcutDialog(props: Props) {
                     />
                 </div>
 
-                <Button
-                    onClick={() => {
-                        validate();
-                    }}
-                >
-                    {validated ? "Add" : "Fetch Info"}
-                </Button>
+                <Spacer height={16} />
+                {error && (
+                    <>
+                        <Alert message={error} />
+                        <Spacer height={32} />
+                    </>
+                )}
+                <div className="actions">
+                    <Button
+                        onClick={() => {
+                            validate();
+                        }}
+                    >
+                        {validated ? "Add" : "Fetch Info"}
+                        {validating && !validated && " ..."}
+                    </Button>
+                    <Button
+                        onClick={() => {
+                            // Reset
+                            setIcon("");
+                            setTitle("");
+                            setUrl("");
+                            setValidated(false);
+                            setValidating(false);
+                            setError(undefined);
+
+                            // Close dialog
+                            props.onClose();
+                        }}
+                    >
+                        Close
+                    </Button>
+                </div>
             </form>
         </div>
     );